+++ source/network/minecraft/StartGame.cpp
@@ -12,7 +12,7 @@ StartGame::StartGame(float x, float y, float z, int gamemode, int seed, int eid)
 }
 
 void StartGame::encode() {
-	this->putByte(MinecraftPackets::START_GAME);
+	this->putByte((uint8)MinecraftPackets::START_GAME);
 	this->putInt(seed);
 	this->putInt(unknown);
 	this->putInt(gamemode);
